Remove the media query
Bug in the media query due to what appears to be an incompatibility
between the min-height property and the implementation of Flexbox by
browsers, resulting in various bugs. Managed to have a correct (even
even though sub-optimal) output with Firefox before realizing that
the solution did not work under Chrome. Have thus to submit a
non-responsive work, sorry.